requirements.txt
Django==6.0.1
djangorestframework==3.16.1
requests==2.32.5
mysqlclient==2.2.7